+++ chift/models/consumers/commerce/__init__.py
@@ -1,10 +1,11 @@
 from typing import ClassVar
 
-from chift.api.mixins import CreateMixin, PaginationMixin, ReadMixin
+from chift.api.mixins import PaginationMixin, ReadMixin
 from chift.openapi.models import CommerceCustomer as CommerceCustomerModel
 from chift.openapi.models import CommerceLocation as CommerceLocationModel
 from chift.openapi.models import CommerceOrder as CommerceOrderModel
 from chift.openapi.models import CommerceProduct as CommerceProductModel
+from chift.openapi.models import CommerceVariant as CommerceVariantModel
 
 
 class CommerceRouter:
@@ -13,6 +14,7 @@ class CommerceRouter:
         self.Product = Product(consumer_id, connection_id)
         self.Location = Location(consumer_id, connection_id)
         self.Order = Order(consumer_id, connection_id)
+        self.Variant = Variant(consumer_id, connection_id)
 
 
 class Customer(
@@ -31,6 +33,15 @@ class Product(
     chift_vertical: ClassVar = "commerce"
     chift_model: ClassVar = "products"
     model = CommerceProductModel
+
+
+class Variant(
+    ReadMixin[CommerceVariantModel],
+    PaginationMixin[CommerceVariantModel],
+):
+    chift_vertical: ClassVar = "commerce"
+    chift_model: ClassVar = "variants"
+    model = CommerceVariantModel
 
 
 class Location(